- advanced coordinate control for overlays over DTA_Fullscreen images.
Manually merged, there were severe conflicts.

# Conflicts:
#	src/textures/textures.h
#	src/v_draw.cpp
#	src/v_video.h

+++ src/polyrenderer/scene/poly_sky.cpp
@@ -228,7 +228,7 @@ Mat4f PolySkyDome::GLSkyMath()
 
 		modelMatrix = Mat4f::Rotate(-180.0f + x_offset, 0.f, 0.f, 1.f);
 
-		float xscale = texw < 1024.f ? floor(1024.f / float(texw)) : 1.f;
+		float xscale = texw < 1024.f ? floorf(1024.f / float(texw)) : 1.f;
 		float yscale = 1.f;
 		if (texh <= 128 && (level.flags & LEVEL_FORCETILEDSKY))
 		{
